Create the choice class
